fix var name conflict by changing
    vec3 R = refract(I, -N, ratio);
    return textureCube(skybox, R.xzy).rgb;
to
    vec3 _R = refract(I, -N, ratio);
    return textureCube(skybox, _R.xzy).rgb;

as R is a uniform in DE-Raytracer.frag used for closest distance to origin during orbit test

this is an odd bug because initially the fragment fails on load but then compiles
when initializeFragment() is called so it's fine the first time but on subsequent
attemps it fails. This does not show up in the automated load/render/save cycle of
the test, it works as expected in that case.